Avoid using enum for flag comparison in gtest

+++ test/test_bonding.cpp
@@ -767,7 +767,7 @@ TEST(Bonding, ConnectNonBlocking)
                 EXPECT_EQ(ev[0].fd, g_listen_socket);
 
                 // Check if the IN event is set, even if it's not the only event
-                EXPECT_EQ(ev[0].events & SRT_EPOLL_IN, SRT_EPOLL_IN);
+                EXPECT_EQ(ev[0].events & SRT_EPOLL_IN, (int)SRT_EPOLL_IN);
                 bool have_also_update = ev[0].events & SRT_EPOLL_UPDATE;
 
                 sockaddr_any adr;
@@ -788,7 +788,7 @@ TEST(Bonding, ConnectNonBlocking)
                     uwait_res = srt_epoll_uwait(lsn_eid, ev, 3, 5000);
                     EXPECT_EQ(uwait_res, 1);
                     EXPECT_EQ(ev[0].fd, g_listen_socket);
-                    EXPECT_EQ(ev[0].events, SRT_EPOLL_UPDATE);
+                    EXPECT_EQ(ev[0].events, (int)SRT_EPOLL_UPDATE);
                 }
 
                 cout << "[A] Waitig for close (up to 5s)\n";
